Update #16 Merging metric and rules analyzer
The analyzer service is now only running only one analysis with check
method. This method is now being run two times in the UI plugin, one for
the metric and a second time for the rules. MetricsViews was updated to
take it in consideration and xml export nows add resultValue attribute
when necessary.
Test were modified to also work with theses modifications