Fix #58 Update parameters in New File CreationPage with default values
Initial content was reaching a null pointer when trying to retrieve the
WizardNewFileCreationPage parameters map because it wasn't updated with
default values (retrieved from Export service).
The export wizard page now update the parameters list of the
WizardNewFileCreationPage not only when a field is modified but also on
creation with default values.