Fix #50 & Update #77
- Every exceptions thrown by checkers are sending more informative
messages

- Code was improved as defined in #77 with a logger & part of PMD &
checkstyle issues were solved

About the logger, a new class UILogger was made. It runs
org.eclipse.logger logger so all the logs are written directly to /*.log
file of ./metadata of workspace and /configuration of eclipse folder.
To precise the level of log a parameter must be indicated on the launch,
as defined in this page :
(https://help.eclipse.org/neon/index.jsp?topic=%2Forg.eclipse.platform.doc.isv%2Freference%2Fmisc%2Fruntime-options.html
). So the following argument can be used to define level of the logger
-Declipse.log.level=<level> where level can be INFO, WARNING, or ERROR.

An important change was made on ConfigurationPagePreferences that was
using CheckerTableViewer and MetricsTableViewer in a composite
Container. These class are now extending Composite that contains a
search SWT.label, a SWT.Text (for the search input) and a TableViewer
with Checkers and parameters relatives to them. Old composite containing
these class was replaced by the new CheckerComposite and
MetricsComposite.